requirements.txt
opencv-python==4.4.0.46
Pillow==9.2.0
blendmodes==2020